Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-import-PBKDF2.txt

imported key: [object CryptoKey]
imported key.type: secret
imported key.extractable: false
imported key.algorithm: {"name":"PBKDF2"}
imported key.usages: deriveKey,deriveBits
SHA-256 digest: 315f5bdb76d078c43b8ac0064e4a0164612b1fce77c869345bfc94c75894edd3
